Misc
* Removed the _TP14 directories.
* Finished the WaterViewer component.
* Construction can now be done on 'isSpace' tiles, but not space itself.
