+++ Content.Client/_TP/Guidebook/Controls/GuideDeepFryerEmbed.xaml
