+++ Content.Client/_TP/HandCrank/HandCrankRechargerSystem.cs
@@ -1,7 +1,8 @@
+using Content.Shared._TP.HandCrank;
 using Content.Shared.Weapons.Ranged.Components;
 using Content.Shared.Weapons.Ranged.Systems;
 
-namespace Content.Client.Weapons.Ranged.Systems;
+namespace Content.Client._TP.HandCrank;
 
 public partial class HandCrankRechargerSystem : SharedHandCrankRechargerSystem
 {
