+++ Content.Client/_TP/Kitchen/DeepFriedSystem.cs
@@ -1,23 +1,24 @@
+using Content.Shared._TP.Kitchen.Components;
 using Content.Shared._TP.Kitchen.Events;
 using Robust.Client.GameObjects;
 using Robust.Shared.Containers;
 
-namespace Content.Client._TP14.Kitchen;
+namespace Content.Client._TP.Kitchen;
 
 /// <summary>
 ///     Client-side sprite visualizer system for the deep-fried item component.
 ///     Created by Cookie (FatherCheese) for Trieste Port 14.
 /// </summary>
-public sealed class DeepFriedVisualizerSystem : EntitySystem
+public sealed class DeepFriedSystem : EntitySystem
 {
     [Dependency] private readonly SpriteSystem _sprite = default!;
 
     public override void Initialize()
     {
         base.Initialize();
-        SubscribeLocalEvent<Shared._TP.Kitchen.Components.SharedDeepFriedComponent, ComponentStartup>(OnComponentStartup);
-        SubscribeLocalEvent<Shared._TP.Kitchen.Components.SharedDeepFriedComponent, DeepFriedLevelChangedEvent>(OnFriedLevelChanged);
-        SubscribeLocalEvent<Shared._TP.Kitchen.Components.SharedDeepFriedComponent, EntGotRemovedFromContainerMessage>(OnRemovedFromContainer);
+        SubscribeLocalEvent<SharedDeepFriedComponent, ComponentStartup>(OnComponentStartup);
+        SubscribeLocalEvent<SharedDeepFriedComponent, DeepFriedLevelChangedEvent>(OnFriedLevelChanged);
+        SubscribeLocalEvent<SharedDeepFriedComponent, EntGotRemovedFromContainerMessage>(OnRemovedFromContainer);
     }
 
     /// <summary>
@@ -25,7 +26,7 @@ public sealed class DeepFriedVisualizerSystem : EntitySystem
     /// </summary>
     /// <param name="friedEnt"></param>
     /// <param name="args"></param>
-    private void OnRemovedFromContainer(Entity<Shared._TP.Kitchen.Components.SharedDeepFriedComponent> friedEnt, ref EntGotRemovedFromContainerMessage args)
+    private void OnRemovedFromContainer(Entity<SharedDeepFriedComponent> friedEnt, ref EntGotRemovedFromContainerMessage args)
     {
         UpdateSprite(friedEnt.Owner, friedEnt.Comp);
     }
@@ -35,7 +36,7 @@ public sealed class DeepFriedVisualizerSystem : EntitySystem
     /// </summary>
     /// <param name="friedEnt">SharedDeepFriedComponent entity</param>
     /// <param name="args">DeepFriedLevelChangedEvent arguments</param>
-    private void OnFriedLevelChanged(Entity<Shared._TP.Kitchen.Components.SharedDeepFriedComponent> friedEnt, ref DeepFriedLevelChangedEvent args)
+    private void OnFriedLevelChanged(Entity<SharedDeepFriedComponent> friedEnt, ref DeepFriedLevelChangedEvent args)
     {
         UpdateSprite(friedEnt.Owner, friedEnt.Comp);
     }
@@ -46,7 +47,7 @@ public sealed class DeepFriedVisualizerSystem : EntitySystem
     /// <param name="friedEnt">SharedDeepFriedComponent entity</param>
     /// <param name="args">ComponentStartup arguments</param>
     /// <exception cref="NotImplementedException"></exception>
-    private void OnComponentStartup(Entity<Shared._TP.Kitchen.Components.SharedDeepFriedComponent> friedEnt, ref ComponentStartup args)
+    private void OnComponentStartup(Entity<SharedDeepFriedComponent> friedEnt, ref ComponentStartup args)
     {
         UpdateSprite(friedEnt.Owner, friedEnt.Comp);
     }
@@ -56,7 +57,7 @@ public sealed class DeepFriedVisualizerSystem : EntitySystem
     /// </summary>
     /// <param name="friedUid">SharedDeepFriedComponent entity uid</param>
     /// <param name="friedComp">SharedDeepFriedComponent entity</param>
-    private void UpdateSprite(EntityUid friedUid, Shared._TP.Kitchen.Components.SharedDeepFriedComponent friedComp)
+    private void UpdateSprite(EntityUid friedUid, SharedDeepFriedComponent friedComp)
     {
         // Simple enough - we check if we have a sprite (we always should),
         // and then we set the color based on the fried level from the component.
@@ -65,9 +66,9 @@ public sealed class DeepFriedVisualizerSystem : EntitySystem
 
         var color = friedComp.CurrentFriedLevel switch
         {
-            Shared._TP.Kitchen.Components.SharedDeepFriedComponent.FriedLevel.LightlyFried => Color.FromHex("#FFD580"),
-            Shared._TP.Kitchen.Components.SharedDeepFriedComponent.FriedLevel.Fried => Color.FromHex("#954535"),
-            Shared._TP.Kitchen.Components.SharedDeepFriedComponent.FriedLevel.Burnt => Color.FromHex("#0E0504"),
+            SharedDeepFriedComponent.FriedLevel.LightlyFried => Color.FromHex("#FFD580"),
+            SharedDeepFriedComponent.FriedLevel.Fried => Color.FromHex("#954535"),
+            SharedDeepFriedComponent.FriedLevel.Burnt => Color.FromHex("#0E0504"),
             _ => Color.White
         };
