+++ Content.Server/Weapons/Ranged/Systems/HandCrankRechagerSystem.cs
@@ -1,4 +1,5 @@
 using Content.Server.Power.EntitySystems;
+using Content.Shared._TP.HandCrank;
 using Content.Shared.Weapons.Ranged.Components;
 using Robust.Shared.Audio.Systems;
 using Content.Shared.Audio;
