+++ Resources/Prototypes/Maps/trieste.yml
@@ -59,4 +59,3 @@
             ResearchAssistant: [ 1, 2 ]
             #silicon
             Borg: [ 2, 2 ]
-            StationAi: [ 1, 1 ]
